+++ ChangeLog.md
@@ -2,6 +2,11 @@
 
 [TOC]
 
+## v1.1
+
+* add special pairs section pairs/ to HOOMD schema
+* bump HOOMD schema version to 1.1
+
 ## v1.0.1
 
 * Fix compile error on more strict POSIX sytstems.